Use aggregate_add_host in nova-manage
When nova-manage placement sync_aggregates was added [1], it duplicated
some report client logic (aggregate_add_host) to do provider aggregate
retrieval and update so as not to duplicate a call to retrieve the
host's resource provider record. It also left a TODO to handle
generation conflicts.

Here we change the signature of aggregate_add_host to accept *either*
the host name or RP UUID, and refactor the nova-manage placement
sync_aggregates code to use it.

The behavior in terms of exit codes and messaging should be largely
unchanged, though there may be some subtle differences in corner cases.

[1] Iac67b6bf7e46fbac02b9d3cb59efc3c59b9e56c8

Change-Id: Iaa4ddf786ce7d31d2cee660d5196e5e530ec4bd3