Move legacy-grenade-dsvm-neutron-multinode-live-migration in-tree
This moves the legacy-grenade-dsvm-neutron-multinode-live-migration
job from openstack-zuul-jobs in-tree. Nova is the only project
that runs this job and this allows nova to directly manage it.

The job will still use the same irrelevant-files list and is still
non-voting for the time being and only runs in the check queue.

It has been renamed to "nova-grenade-live-migration".

This change is meant to be backported to rocky, queens and pike
to allow us to drop legacy-grenade-dsvm-neutron-multinode-live-migration
from openstack-zuul-jobs where the job is branchless but restricted
to pike+ branches.

A follow up master-only (stein) change will make the job voting
and gate per the plan in the mailing list [1].

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-February/003341.html

Change-Id: Ie9b61775dbb92b10237688eaddaca606c1c73a23