fakelibvirt: Add ability to generate fake PCI devices
We already support generating fake PFs and VFs. Now add the ability to
generate fake non-SR-IOV devices. This is mostly an exercise in
refactoring, as we remove a lot of unused configurability and generally
clean things up as we go.

This isn't used yet. We do that in a later patch.

Change-Id: I22e3166ef3b706add88c6b909af024022f2243a8