Improve existing flavor and image metadata validation
This improves the existing validation of flavor extra-specs and image
properties in hardware.py in preparation for calling the validation from
more places in a follow-on patch.

get_cpu_topology_constraints() becomes public as we'll need to call it
from the API code in the next patch in the series.

If the CPU topology is not valid, raise an InvalidRequest exception with
a useful error message instead of a ValueError message without any
context.

Add checks to ensure that the CPU and CPU thread policies are valid,
and if not then raise newly-added exceptions with useful error messages.

Tweak various docstrings to more accurately reflect what exceptions
they might raise.

Change-Id: I20854134c80b8f4598f375eae137fd2920114891
blueprint: flavor-extra-spec-image-property-validation
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>