Use a placement conf when testing report client
It turns out that the independent wsgi interceptors in
test_report_client were using nova's global configuration
when creating the intercepts using code from placement.
This was working because until [1] placement's set of conf
options had not diverged from nova's and nova still has
placement_database config settings.

This change takes advantage of new functionality in the
PlacementFixture to allow the fixtur to manage config and
database, but _not_ run the interceptor. This means it
can set up a config that is later used by the independent
interceptors that are used in the report client tests.

[1] Ie43a69be8b75250d9deca6a911eda7b722ef8648

Change-Id: I05326e0f917ca1b9a6ef8d3bd463f68bd00e217e
Closes-Bug: #1818560
Depends-On: I8c36f35dbe85b0c0db1a5b6b5389b160b68ca488