Refactor pre-live-migration work out of _do_live_migration
The _do_live_migration method has gotten too big so this
helps by refactoring the pre-live-migration portion into
a new private method with documentation. Note that because
of how _do_live_migration was shadowing and overwriting the
migrate_data input argument the new method has to pass that
back as well. Failing to do that causes integration tests
to fail because unit tests were insufficient, so this also
updates one unit test to make sure that works properly.

This is a small part of a bigger long-term effort to try
and cleanup these live migration compute manager methods
so they can be maintained and tested in isolation, and also
reason about error handling for each piece.

Change-Id: I220a4f9337d355a76596208cf7b96113dd34a8ea