Clean up block_device_allocate_retries config option help
The "on failures" portion of this option's description does
not make sense. It's not about failures, it's about how many
times to check a volume that nova-compute created to be
"available" before it can be attached to the server during
boot from volume. So that is replaced and a new description is
added explaining when and how this is used, and what happens
on timeout.

The old Liberty mention of the volume image cache is also replaced
with a link to the cinder docs for that feature.

Change-Id: Ic7ef9d44ec64c3d5834181bd8b61aae56858af81
Partial-Bug: #1819049