Refactor ComputeManager.remove_volume_connection
This change does two things:

First is a simple refactor to allow passing a BDM to
_remove_volume_connection. The interface on the public
remove_volume_connection cannot be easily changed since
it's an RPC API call.

Second, the new private method gains a delete_attachment
kwarg which will tell the method whether or not to delete
a v3-style attachment, i.e. bdm.attachment_id is not None.

This is prep work for an upcoming change that is going
to call the new private method locally rather than over
RPC.

Change-Id: Ice6d5b4757134b7d1b36514744a8f958528a6a47