fix race in test_interface_detach_with_port_with_bandwidth_request
During port detach the unbind towards neutron happens before the
port allocation is removed from placement. The functional test only
waited for the port unbind before asserted the remaining allocations and
therefore it was racy.

Fortunately the instance.interface_detach.end is emitted after the both
the unbind and the allocation shrink. So the test is changed to wait for
this notification instead.

Change-Id: I53d76d6353ae634e387672e14943f518955b221e
Closes-Bug: #1819374