Validate image/create during cross-cell resize functional testing
During a cross-cell resize, image-backed servers will have a
snapshot image created from the source cell instance and that
will be used to spawn the guest in the destination host. This
adds validation to the functional tests to ensure a snapshot
image is created during the resize for image-backed servers and
is also cleaned up when the resize is finished. There is no
snapshot created for volume-backed servers so for those types
of servers the test just asserts that no snapshot image was
created.

Part of blueprint cross-cell-resize

Change-Id: I5b59eeb8c29d3c6f13326ed2ce29212995d105ea