Skip the ceph based live migration testing
While moving the legacy jobs on Bionic[1], it is found
that ceph pckg hammer is not available to download
on bionic and it error with 404. Due to that job testing the
live migration on ceph not working:
-  nova-live-migration skip the ceph live migration testing[2]
-  nova-grenade-live-migration fail due to no ceph pckg[3]

To proceed further on Bionic migration which is blocked by these failure,
this commit temporary skip the ceph live migration testing until
bug#1819944 is fixed.

Related-Bug: 1819944
[1] https://review.openstack.org/#/c/639017
[2] http://logs.openstack.org/17/639017/8/check/nova-live-migration/87084aa/logs/devstack-gate-post_test_hook.txt.gz#_2019-03-13_14_34_20_967
[3] http://logs.openstack.org/17/639017/8/check/nova-grenade-live-migration/735242b/logs/devstack-gate-post_test_hook.txt.gz#_2019-03-13_15_00_32_310

Change-Id: I05182d8fd0df5e8f3f9f4fb11feed074990cdb9f