Add functional recreate test for bug 1819963
When resizing a server that was not created in an explicit
zone, the scheduler can pick a host in another zone and
conductor will update the instance.availability_zone value
for the new dest host zone.

The problem is when reverting the resize, the server goes
back to the original source host/zone but the
instance.availability_zone value in the database is not
updated which can lead to incorrect results when listing
servers and filtering by zone.

This adds a functional recreate test for the bug.

Change-Id: Ib107650d6a2c991c26b646a0dd10ddc7a3fb7e56
Related-Bug: #1819963
(cherry picked from commit 73aaead294b9df412305abb1cb01aac95477bcc1)