Add zones wrinkle to TestMultiCellMigrate
This creates an availability zone for each cell in
the functional test to ensure that the instance
availability_zone in the database is properly updated
when migrating the instance across cells and zones.

This can also be used later during revert resize testing
to ensure the instance AZ is properly updated on revert.

Part of blueprint cross-cell-resize

Change-Id: I97d0567c77eaaa9af0895ed58918c4c61377e926