Add negative test for cross-cell finish_resize failing
This adds a functional test where the finish resize on the
dest compute fails triggering a rollback in conductor. The
test also hard reboots the server in the source cell to
recover it from ERROR to ACTIVE status.

Finally, the server is resized again without the fault
injected to make sure the target cell DB was properly
cleaned up on rollback from the initial resize failure.

Because prep_snapshot_based_resize_at_source will destroy
the guest from the source hypervisor, this change needs to
implement some reboot logic in the fake driver such that
if a hard reboot is being attempted on a guest that is not
being tracked, the fake driver will (re)spawn the guest
similar to how hard reboot works for the libvirt driver.

Part of blueprint cross-cell-resize

Change-Id: If6a8cf58944366917a729527cb3b87494de38fd0