Remove unnecessary default provider_tree when getting traits
I15364d37fb7426f4eec00ca4eaf99bec50e964b6 introduced a new
_get_traits() method in ResourceTracker for getting traits from a
provider tree and then merging with capability traits from the driver.
However it included a default of None for the provider_tree parameter
which was a remnant of earlier iterations of that change.  Since this
method is always passed a ProviderTree, remove the superfluous default
of None.

Change-Id: I1868485912d9a8a330bde50836808accf04c728d